Updating DSTextButton
